Also check if saver is permitted user for m.NOTICE
Previously we would only have checked for m.MESSAGE. As an aside, we
only check for is_text() events now, so technically this allows saving
an emote as well. There is no reason why we should exclude those...

I have to admit I don't understand what the bot check in m.NOTICE checks
is supposed to do or what it protects against, so I just kept it in.